Improve job id extraction
Some versions of the Unix batch command produce extra lines of output on
stderr, so the original assumption that the jobid is the second word of
output is not always correct. Replaced this with a regexp that ignores
arbitrary lines of output on either side of the "job N" line.